extract components from application
