+++ lib/components/active-file.js
@@ -2,11 +2,9 @@ import React from 'react';
 import sendGist from '../send-gist';
 import MarkdownButtons from './markdown-buttons';
 import MarkdownBox from './markdown-box';
-import updateContent from '../update-content'
+import updateContent from '../update-content';
 
-
-const ActiveFile = ( {file} ) => {
-
+module.exports = ( {file} ) => {
 return (
     <div className="active-file">
       <MarkdownButtons {...file}/>
@@ -15,5 +13,3 @@ return (
     </div>
   );
 };
-
-module.exports = ActiveFile;
