+++ lib/components/application.js
@@ -3,15 +3,10 @@ import store from '../store';
 import FileList from './file-list';
 import { remote } from 'electron';
 import ActiveFile from './active-file';
-const Selfup = require('selfup-rejs');
-const openFile = remote.require(`${__dirname}/../open-file`);
-const { saveFile } = remote.require(`${__dirname}/../main`);
 const lastActiveFile = require('../last-active-file');
-const rejs = new Selfup();
 import saveOnClose from '../save-on-close';
 import loadSavedFiles from '../load-saved-files';
-
-const currentWindow = remote.getCurrentWindow();
+import Controls from './controls';
 
 saveOnClose();
 loadSavedFiles();
@@ -33,13 +28,7 @@ class Application extends Component {
 
     return (
       <div>
-        <div className="controls">
-          <button className="controls-new-file" onClick={() => store.newFile()}>New Note</button>
-          <button className="controls-open-file" onClick={() => openFile(currentWindow)}>Open File</button>
-          {saveFileButton(activeFile)}
-          <button className="controls-open-file" onClick={() => store.deselect()}>Close current file</button>
-          {renderTitle(activeFile)}
-        </div>
+        <Controls {...activeFile}/>
         <div className="files">
           <FileList files={this.state.files} />
           <ActiveFile file={activeFile} />
@@ -48,19 +37,5 @@ class Application extends Component {
     );
   }
 }
-
-const renderTitle = (file) => {
-  if (file) {
-    return (
-      <h3>{file.fileName}</h3>
-    );
-  }
-};
-
-const saveFileButton = (activeFile) => {
-  if (activeFile) {
-    return <button className="controls-open-file" onClick={() => saveFile(activeFile.fileName, activeFile.content, activeFile.id)}>Save Current File</button>;
-  }
-};
 
 export default Application;
